(split)bug: 30550
Blob size is a long, now matching MailboxBlob.

http://bugzilla.zimbra.com/show_bug.cgi?id=30550

Copied from Perforce
 Change: 165722

+++ src/java/com/zimbra/cs/mime/ParsedDocument.java
@@ -62,7 +62,7 @@ public class ParsedDocument {
     public ParsedDocument(Blob blob, String filename, String ctype, long createdDate, String creator)
     throws ServiceException, IOException {
         mBlob = blob;
-        mSize = blob.getRawSize();
+        mSize = (int) blob.getRawSize();
         mDigest = blob.getDigest();
         mContentType = ctype;
         mFilename = filename;
